requirements.txt
aiogram
aiosqlite
aiohttp-socks
psycopg
